put settings inside app
