+++ client/src/App.tsx
@@ -6,7 +6,7 @@ import CssBaseline from "@mui/material/CssBaseline";
 import { saveColorTheme } from "./state/persistance";
 import type { ColorTheme } from "./styles/modes";
 import { themes } from "./styles/modes";
-
+import AppBar from "./components/AppBar";
 import { SetThemeEnum, useThemeParams } from "./state/themeActions";
 import { Outlet } from "react-router";
 
@@ -16,12 +16,18 @@ function App() {
     themeDispatch({ type: SetThemeEnum.UPDATE, value: theme });
     saveColorTheme(theme);
   };
+  const hasAuthSet = sessionStorage.getItem("admin_password") ? true : false;
   const theme = themes[selectedTheme];
   return (
     <ThemeProvider theme={theme}>
       <Box sx={{ display: "flex" }}>
         <CssBaseline />
-        <Outlet context={{ selectedTheme, setThemeAndSave }} />
+        <AppBar
+          mode={selectedTheme}
+          setMode={setThemeAndSave}
+          isAdmin={hasAuthSet}
+        />
+        <Outlet />
       </Box>
     </ThemeProvider>
   );
