+++ client/src/components/AppBar.tsx
@@ -10,7 +10,7 @@ import ToggleButton from "@mui/material/ToggleButton";
 import IconButton from "@mui/material/IconButton";
 import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import type { ColorTheme } from "../styles/modes";
-import { Link } from "react-router";
+import { NavLink } from "react-router";
 
 const drawerWidth: number = 240;
 interface AppBarProps extends MuiAppBarProps {
@@ -48,16 +48,18 @@ const AppBarMenu = ({ setMode, mode, isAdmin }: AppBarMenuProps) => {
         }}
       >
         <Typography
-          component="h1"
+          component={NavLink}
+          to="/"
           variant="h6"
           color="inherit"
           noWrap
           sx={{ flexGrow: 1 }}
+          style={{ textDecoration: "none", color: "inherit" }}
         >
           MiniDSP
         </Typography>
         {isAdmin && (
-          <IconButton component={Link} to="/settings" aria-label="settings">
+          <IconButton component={NavLink} to="/settings" aria-label="settings">
             <SettingsIcon />
           </IconButton>
         )}
